refactor(SCB-1201): move Image to image/, fix map imports, deduplicate site setup scripts
