+++ examples/kit-nextjs-article-starter/src/__tests__/components/image/Image.mockProps.ts
