+++ examples/kit-nextjs-article-starter/src/__tests__/components/image/Image.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { Default as Image, Banner } from '@/components/sxa/Image';
+import { Default as Image, Banner } from '@/components/image/Image';
 import type { ImageProps } from '@/components/image/image.props';
 import type { ImageField } from '@sitecore-content-sdk/nextjs';
 import {
